requirements.txt
streamlit
google-generativeai
pandas
matplotlib
plotly
Pillow